// ui/src/generated/BoardSpaceCategories.ts
/**
 * This is a TypeGen auto-generated file.
 * Any changes made to this file can be lost when this file is regenerated.
 */

export enum BoardSpaceCategories {
    Null = 0,
    Go = 1,
    Property = 2,
    Railroard = 3,
    Utility = 4,
    Jail = 5,
    FreeParking = 6,
    GoToJail = 7,
    PayTaxes = 8,
    Chance = 9,
    CommunityChest = 10,
}